@@ -1,6 +1,6 @@
 {
   "name": "@nestjs/core",
-  "version": "7.5.3",
+  "version": "7.6.8",
   "description": "Modern, fast, powerful node.js web framework",
   "homepage": "https://nestjs.com",
   "repository": {
@@ -116,7 +116,7 @@
     "coveralls": "3.1.0",
     "delete-empty": "3.0.0",
     "engine.io-client": "4.1.0",
-    "eslint": "7.18.0",
+    "eslint": "7.19.0",
     "eslint-config-prettier": "7.2.0",
     "eslint-plugin-import": "2.22.1",
     "eventsource": "1.0.7",
@@ -136,7 +136,7 @@
     "gulp-watch": "5.0.1",
     "http-errors": "1.8.0",
     "husky": "4.3.8",
-    "imports-loader": "1.2.0",
+    "imports-loader": "2.0.0",
     "json-loader": "0.5.7",
     "kafkajs": "1.15.0",
     "lerna": "2.11.0",
@@ -146,14 +146,14 @@
     "merge-graphql-schemas": "1.7.8",
     "middie": "5.2.0",
     "mocha": "8.2.1",
-    "mongoose": "5.11.13",
+    "mongoose": "5.11.14",
     "mqtt": "4.2.6",
     "multer": "1.4.2",
     "mysql": "2.18.1",
     "nats": "1.4.12",
     "nodemon": "2.0.7",
     "nyc": "15.1.0",
-    "point-of-view": "4.9.0",
+    "point-of-view": "4.11.0",
     "prettier": "2.2.1",
     "redis": "3.0.2",
     "rxjs-compat": "6.6.3",
