@@ -42,13 +42,13 @@
     "jest": "26.6.3",
     "prettier": "2.2.1",
     "supertest": "6.1.3",
-    "ts-jest": "26.4.4",
+    "ts-jest": "26.5.0",
     "ts-loader": "8.0.14",
     "ts-node": "9.1.1",
     "tsconfig-paths": "3.9.0",
     "@typescript-eslint/eslint-plugin": "4.14.1",
     "@typescript-eslint/parser": "4.14.1",
-    "eslint": "7.18.0",
+    "eslint": "7.19.0",
     "eslint-config-prettier": "7.2.0",
     "eslint-plugin-import": "2.22.1",
     "typescript": "4.1.3"
