@@ -23,7 +23,7 @@
     "@nestjs/core": "7.6.7",
     "@nestjs/mongoose": "7.2.2",
     "@nestjs/platform-express": "7.6.7",
-    "mongoose": "5.11.13",
+    "mongoose": "5.11.14",
     "reflect-metadata": "0.1.13",
     "rimraf": "3.0.2",
     "rxjs": "6.6.3"
@@ -40,13 +40,13 @@
     "jest": "26.6.3",
     "prettier": "2.2.1",
     "supertest": "6.1.3",
-    "ts-jest": "26.4.4",
+    "ts-jest": "26.5.0",
     "ts-loader": "8.0.14",
     "ts-node": "9.1.1",
     "tsconfig-paths": "3.9.0",
     "@typescript-eslint/eslint-plugin": "4.14.1",
     "@typescript-eslint/parser": "4.14.1",
-    "eslint": "7.18.0",
+    "eslint": "7.19.0",
     "eslint-config-prettier": "7.2.0",
     "eslint-plugin-import": "2.22.1",
     "typescript": "4.1.3"
