@@ -12,15 +12,17 @@ import {
   EventPattern,
   MessagePattern,
   Transport,
+  RpcException,
 } from '@nestjs/microservices';
-import { from, Observable, of } from 'rxjs';
-import { scan } from 'rxjs/operators';
+import { from, Observable, of, throwError } from 'rxjs';
+import { catchError, scan } from 'rxjs/operators';
 
 @Controller()
 export class AppController {
   constructor(
     @Inject('USE_CLASS_CLIENT') private useClassClient: ClientProxy,
     @Inject('USE_FACTORY_CLIENT') private useFactoryClient: ClientProxy,
+    @Inject('CUSTOM_PROXY_CLIENT') private customClient: ClientProxy,
   ) {}
   static IS_NOTIFIED = false;
 
@@ -75,6 +77,17 @@ export class AppController {
       .reduce(async (a, b) => (await a) && b);
   }
 
+  @Post('error')
+  @HttpCode(200)
+  serializeError(@Query('client') query: 'custom' | 'standard' = 'standard', @Body() body: Record<string, any>): Observable<boolean> {
+    const client = query === 'custom' ? this.customClient : this.client;
+    return client.send({ cmd: 'err' }, {}).pipe(
+      catchError((err) => {
+        return of(err instanceof RpcException);
+      })
+    )
+  }
+
   @MessagePattern({ cmd: 'sum' })
   sum(data: number[]): number {
     return (data || []).reduce((a, b) => a + b);
@@ -93,6 +106,11 @@ export class AppController {
   @MessagePattern({ cmd: 'streaming' })
   streaming(data: number[]): Observable<number> {
     return from(data);
+  }
+
+  @MessagePattern({ cmd: 'err' })
+  throwAnError() {
+    return throwError(new Error('err'));
   }
 
   @Post('notify')
