@@ -1,13 +1,17 @@
 import { RpcParamtype } from '../enums/rpc-paramtype.enum';
 
 export class RpcParamsFactory {
-  public exchangeKeyForValue(type: number, args: unknown[]) {
+  public exchangeKeyForValue(
+    type: number,
+    data: string | undefined,
+    args: unknown[],
+  ) {
     if (!args) {
       return null;
     }
     switch (type as RpcParamtype) {
       case RpcParamtype.PAYLOAD:
-        return args[0];
+        return data ? args[0]?.[data] : args[0];
       case RpcParamtype.CONTEXT:
         return args[1];
       case RpcParamtype.GRPC_CALL:
