@@ -1,32 +1,40 @@
+
 /** ------------------------------------------------------
  * THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
  * -------------------------------------------------------
  */
 
 /* tslint:disable */
-export abstract class CreateCatInput {
-  name?: string;
-  age?: number;
-}
-
-export abstract class Cat {
-  id?: number;
-  name?: string;
-  age?: number;
-}
-
-export abstract class IMutation {
-  abstract createCat(createCatInput?: CreateCatInput): Cat | Promise<Cat>;
+/* eslint-disable */
+export class CreateCatInput {
+    name?: string;
+    age?: number;
 }
 
 export abstract class IQuery {
-  abstract getCats(): Cat[] | Promise<Cat[]>;
+    abstract cats(): Cat[] | Promise<Cat[]>;
 
-  abstract cat(id: string): Cat | Promise<Cat>;
+    abstract cat(id: string): Cat | Promise<Cat>;
+}
 
-  abstract temp__(): boolean | Promise<boolean>;
+export abstract class IMutation {
+    abstract createCat(createCatInput?: CreateCatInput): Cat | Promise<Cat>;
 }
 
 export abstract class ISubscription {
-  abstract catCreated(): Cat | Promise<Cat>;
+    abstract catCreated(): Cat | Promise<Cat>;
+}
+
+export class Owner {
+    id: number;
+    name: string;
+    age?: number;
+    cats?: Cat[];
+}
+
+export class Cat {
+    id?: number;
+    name?: string;
+    age?: number;
+    owner?: Owner;
 }
